# src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Babeonym API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "db_error",
              "server_error"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "number",
            "example": 123
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "authProvider": {
            "type": "string",
            "nullable": true,
            "description": "AuthProvider enum value"
          },
          "theme": {
            "type": "string",
            "nullable": true
          },
          "surName": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "GivenName": {
        "type": "object",
        "required": [
          "givenName",
          "givenCustomNameBridgeId",
          "rating"
        ],
        "properties": {
          "givenName": {
            "type": "string",
            "example": "Oliver"
          },
          "givenCustomNameBridgeId": {
            "type": "number",
            "example": 42
          },
          "rating": {
            "type": "number",
            "example": 1000
          },
          "percentile": {
            "type": "number",
            "nullable": true,
            "example": 92
          },
          "gender": {
            "type": "string",
            "nullable": true,
            "description": "Key of Gender enum",
            "example": "male"
          }
        }
      },
      "NotAuthenticatedResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Not authenticated"
          }
        }
      },
...
      "EtymologyLanguage": {
        "type": "object",
        "required": [
          "id",
          "label",
          "flag"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "example": 4
          },
          "label": {
            "type": "string",
            "example": "Arabic"
          },
          "flag": {
            "type": "string",
            "nullable": true,
            "example": "🇸🇦"
          }
        }
      },
      "EtymologyCulture": {
        "type": "object",
        "required": [
          "id",
          "label"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "example": 26
          },
          "label": {
            "type": "string",
            "example": "Persian"
          }
        }
      },
      "EtymologyMeaning": {
        "type": "object",
        "required": [
          "short",
          "long",
          "dateCreated",
          "dateUpdated"
        ],
        "properties": {
          "short": {
            "type": "string",
            "nullable": true,
            "example": "knowledgeable, wise"
          },
          "long": {
            "type": "string",
            "nullable": true,
            "example": "Derived from Arabic meaning one who knows or is aware."
          },
          "dateCreated": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-02-22T14:12:00Z"
          },
          "dateUpdated": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-02-22T14:15:00Z"
          }
        }
      },
      "Etymology": {
        "type": "object",
        "required": [
          "givenNameId",
          "givenName",
          "languages",
          "cultures",
          "meaning"
        ],
        "properties": {
          "givenNameId": {
            "type": "integer",
            "example": 123
          },
          "givenName": {
            "type": "string",
            "example": "Arif"
          },
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EtymologyLanguage"
            }
          },
          "cultures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EtymologyCulture"
            }
          },
          "meaning": {
            "$ref": "#/components/schemas/EtymologyMeaning"
          }
        }
      }
    },
    "responses": {
...
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "paths": {
    "/api/v1/auth/anonymous": {
      "get": {
        "operationId": "v1AuthAnonymous",
        "summary": "Create an anonymous user session",
        "description": "Creates a new anonymous user and logs them in via session cookie.",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Anonymous session created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message",
                    "user"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Anonymous session created"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "User is already authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "User is already authenticated"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/v1/auth/google/callback": {
      "get": {
        "operationId": "v2AuthGoogleCallback",
        "summary": "Handle Google OAuth callback",
        "description": "Handles the Google OAuth callback, links or logs in a user as needed, and redirects to the frontend with success or error state.\n",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "OAuth authorization code returned by Google"
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "OAuth error returned by Google"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to frontend (success, linking flow, or error)"
          }
        }
      }
    },
...
    "/api/v1/auth/microsoft/callback": {
      "get": {
        "operationId": "v1AuthMicrosoftCallback",
        "summary": "Handle Microsoft OAuth callback",
        "description": "Handles the Microsoft OAuth callback, links or logs in a user as needed, and redirects to the frontend with success or error state.\n",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "OAuth authorization code returned by Microsoft"
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "OAuth error returned by Microsoft"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to frontend (success, linking flow, or error)"
          }
        }
      }
    },
...
    "/api/v1/givenName/etymology/{givenCustomNameBridgeId}": {
      "get": {
        "operationId": "v1GivenNameEtymology",
        "summary": "Get given name etymology",
        "description": "Returns etymology information for a specific given name.",
        "tags": [
          "GivenName"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "givenCustomNameBridgeId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the given name"
          }
        ],
        "responses": {
          "200": {
            "description": "Etymology data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Etymology"
                }
              }
            }
          },
          "400": {
            "description": "Invalid givenCustomNameBridgeId",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid givenCustomNameBridgeId"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotAuthenticatedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/givenName/search": {
      "get": {
        "operationId": "v1GivenNameSearch",
        "summary": "Search given names",
        "description": "Searches given names for the authenticated user using a text query.\n",
        "tags": [
          "GivenName"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Search text for given names"
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Maximum number of results to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Matching given names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GivenName"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required search query",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Query parameter 'search' is required"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotAuthenticatedResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}